requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
async-google-trans-new==1.4.1
async-timeout==3.0.1
attrs==21.2.0
brotlipy==0.7.0
cchardet==2.1.7
cffi==1.14.5
chardet==4.0.0
debtcollector==2.2.0
discord.py==1.7.3
idna==3.2
multidict==5.1.0
pbr==5.6.0
pycares==4.0.0
pycparser==2.20
six==1.16.0
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.5
wrapt==1.12.1
yarl==1.6.3
